backend_flask/requirements.txt
Flask>=3.0
python-dotenv>=1.0
PyMySQL>=1.1.0
Flask-SQLAlchemy>=3.1.1

